simple confirmation dialogs for deleting repos and profiles

+++ ModsDude.Client/ModsDude.Client.Wpf/ViewModel/Pages/ProfilePageViewModel.cs
@@ -21,18 +21,23 @@ public partial class ProfilePageViewModel(
     public bool IsValid => !string.IsNullOrWhiteSpace(Name);
 
 
-    [RelayCommand]
-    public async Task DeleteRepo(CancellationToken cancellationToken)
-    {
-        navigationLockService.ReleaseLock(this);
-        await profileService.DeleteProfile(profile.RepoId, profile.Id, cancellationToken);
-    }
-
     [RelayCommand(CanExecute = nameof(IsValid))]
     public async Task SaveChanges(CancellationToken cancellationToken)
     {
         navigationLockService.ReleaseLock(this);
         await profileService.UpdateProfile(profile.RepoId, profile.Id, Name, cancellationToken);
+    }
+
+    [RelayCommand]
+    public async Task DeleteRepo(CancellationToken cancellationToken)
+    {
+        if (!ConfirmDelete())
+        {
+            return;
+        }
+
+        navigationLockService.ReleaseLock(this);
+        await profileService.DeleteProfile(profile.RepoId, profile.Id, cancellationToken);
     }
 
     public void Dispose()
@@ -44,5 +49,18 @@ public partial class ProfilePageViewModel(
     partial void OnNameChanged(string value)
     {
         navigationLockService.AcquireLock(this);
+    }
+
+
+    private bool ConfirmDelete()
+    {
+        var result = System.Windows.Forms.MessageBox.Show(
+            $"Are you sure you want to delete '{OriginalName}'.\nThis action cannot be undone!",
+            "Really?",
+            System.Windows.Forms.MessageBoxButtons.OKCancel,
+            System.Windows.Forms.MessageBoxIcon.Warning,
+            System.Windows.Forms.MessageBoxDefaultButton.Button2);
+
+        return result == System.Windows.Forms.DialogResult.OK;
     }
 }
